requirements.txt
async-timeout==4.0.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
greenlet==3.0.1
idna==3.4
pycparser==2.21
PyMySQL==1.1.0
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.0.7
